+++ projects/clr/hipamd/src/hip_table_interface.cpp
@@ -375,6 +375,14 @@ hipError_t hipDeviceGetName(char* name, int len, hipDevice_t device) {
   return hip::GetHipDispatchTable()->hipDeviceGetName_fn(name, len, device);
   CATCH;
 }
+hipError_t hipDeviceGetP2PAtomicCapabilities(unsigned int* capabilities,
+                                             const hipAtomicOperation* operations,
+                                             unsigned int count, int srcDevice, int dstDevice) {
+  TRY;
+  return hip::GetHipDispatchTable()->hipDeviceGetP2PAtomicCapabilities_fn(
+      capabilities, operations, count, srcDevice, dstDevice);
+  CATCH;
+}
 hipError_t hipDeviceGetP2PAttribute(int* value, hipDeviceP2PAttr attr, int srcDevice,
                                     int dstDevice) {
   TRY;